In the Vulkan API, the timeout parameter of vkWaitForFences is of type uint64_t. If the API is expected to wait indefinitely (until the GPU signals the fence), the correct value to pass is UINT64_MAX (the maximum 64-bit unsigned integer), not the maximum 32-bit value (UINT32_MAX).

+++ framework/fence_pool.h
@@ -1,4 +1,4 @@
-/* Copyright (c) 2019-2020, Arm Limited and Contributors
+/* Copyright (c) 2019-2025, Arm Limited and Contributors
  *
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -40,7 +40,7 @@ class FencePool
 
 	VkFence request_fence();
 
-	VkResult wait(uint32_t timeout = std::numeric_limits<uint32_t>::max()) const;
+	VkResult wait(uint64_t timeout = std::numeric_limits<uint64_t>::max()) const;
 
 	VkResult reset();
